Task 6

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ char b;
 long int c;
 long long int d;
 float f;
-printf("size of a char: %lu bytes\n", sizeof(b));
-printf("size of an int: %lu bytes\n", sizeof(a));
-printf("size of a long int: %lu bytes\n", sizeof(c));
-printf("size of a long long int: %lu bytes\n", sizeof(d));
-printf("size of a float: %lu bytes\n", sizeof(f));
+ printf("size of a char: %lu bytes\n", (unsigned long)sizeof(b));
+ printf("size of an int: %lu bytes\n", (unsigned long) sizeof(a));
+ printf("size of a long int: %lu bytes\n", (unsigned long) sizeof(c));
+ printf("size of a long long int: %lu bytes\n", (unsigned long) sizeof(d));
+ printf("size of a float: %lu bytes\n", (unsigned long) sizeof(f));
 return (0);
 }
